added `ensure_nonzero_selectors` to composer
partial commit - only updated plookup_composer.

when making the proving key, we want
to efficiently ensure that all selector
polynomials are nonzero.

This is so that the verifier can validate that every group element in
the verification key is not the point at infinity.

This makes recursive verification simpler, as the verifier does not
have to conditionally perform a group exponentiation, depending
on whether a given selector polynomial is zero or not.