fix(acir): when retrying failed ACIR tests it should not use the default CLI argument (#1673)
If a test is ran and it fails, the script will try to re-run the test
using verbose mode so that you can debug the failure.

Since we were not passing the flag for the bytecode, it will use the
default path of `./target/main.bytecode` which would inevitably fail
because all of the compiled tests are named after their directories,
like `./target/1_mul.bytecode` .

# Checklist:
Remove the checklist to signal you've completed it. Enable auto-merge if
the PR is ready to merge.
- [ ] If the pull request requires a cryptography review (e.g.
cryptographic algorithm implementations) I have added the 'crypto' tag.
- [ ] I have reviewed my diff in github, line by line and removed
unexpected formatting changes, testing logs, or commented-out code.
- [ ] Every change is related to the PR description.
- [ ] I have
[linked](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
this pull request to relevant issues (if any exist).