chore: Recursion todos (#2516)
Handle some TODOs including:

- (Issue
#[726](https://github.com/AztecProtocol/barretenberg/issues/726))
Reconfigure Sumcheck Verifier to always return a `SumcheckOutput` rather
than a `std::optional<SumcheckOutput>`. This was fine for native
verification but doesn't work in the recursive setting (and the same
SumcheckVerifier code is shared between the two). For native
verification / debugging purposes, the verifier now returns a `verified`
flag as part of `SumcheckOutput`. In the recursive setting, the verifier
does not "abort early" if sumcheck fails, the proof will simply fail to
verify.
- Add failure tests for the Honk recursive verifiers. (There was no
issue associated with this).
- There were several instances of `TODO(luke)` (without a corresponding
issue) that were outdated, easily fixed, or more appropriate as
`Note:`'s