feat: Integrate ZeroMorph into Honk (#2774)
This PR integrates ZeroMorph into Honk. Anywhere we used to use
Gemini+Shplonk+KZG, we now use ZeroMorph.

Note: Charlie has confirmed that the `work_queue` is no longer needed so
it has not been incorporated into the ZeroMorph rounds. (I'll remove the
`work_queue` altogether in a follow on). This also means that it doesn't
make sense to define the ZM protocol across multiple rounds split up in
the Prover/Verifier (like we used to for Gemini/Shplonk). Instead, the
entire protocol is defined via `prove` and `verify` functions in the
`ZeroMorphProver/Verifier` classes and those methods are called from the
main Honk Prover/Verifier (similar to Sumcheck).