chore: Remove unecessary calls to `pedersen__init` (#3079)
Related to #3029 and resolves #3077 .

The generator context is initializing and extending generator data which
outlives the lifetime of the function (it stays alive for the lifetime
of wasm instance). There may be an argument in the future to add a
method that will arbitrarily extend the generators, so that they do not
get extended in the middle of a pedersen call, if this is wanted we can
open up a separate issue.

# Checklist:
Remove the checklist to signal you've completed it. Enable auto-merge if
the PR is ready to merge.
- [ ] If the pull request requires a cryptography review (e.g.
cryptographic algorithm implementations) I have added the 'crypto' tag.
- [ ] I have reviewed my diff in github, line by line and removed
unexpected formatting changes, testing logs, or commented-out code.
- [ ] Every change is related to the PR description.
- [ ] I have
[linked](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
this pull request to relevant issues (if any exist).