refactor: Move transcripts (#3176)
As a a first step toward breaking up the barretenberg/honk submodule,
which is gotten very large, I move the Honk transcript out of Honk. This
will enable moving ZeroMorph and other hings in barretenberg/honk/pcs
out of honk. it is now in barretenberg/transcript. Formerly
barretenberg/transcript was the Plonk transcript. This goes into
barretenberg/plonk.