fix: Fix block constraint key divergence bug. (#3256)
* Fixes the "bad witness key divergence bug" in RAM block constraint by
creating new witness and asserting it's equal to the original witness.
* Adds a new `prove_then_verify` test flow, which is used to always
exercise the good/bad witness data paths to ensure the resulting keys
are the same. This is faster than `all_cmds` which isn't needed for
every test. It's slower than `prove_and_verify` which uses the single
built in command, but that's less safe as it only exercised the bad
witness path for circuit construction.
* Limit hardware concurrency to 16 cores as performance can actually be
worse with higher core counts.