refactor: remove 'extern template's, expand macros (#3953)
- "extern template" was
[cargo-culted](https://en.wikipedia.org/wiki/Cargo_cult_programming).
Deleting it all changes no semantics at all. The feature seems at a
surface level to be used like other uses of C++ 'extern' that indicate
to look in a different compilation unit, such as is used by global
variables in a different translation unit. But, C++ actually already
looks in different translation units by default for all the symbols we
applied it to. The only use of extern template is actually if the source
is provided in a header, the exact opposite of how it was used in the
codebase.
- Simplifies usage of instantiation macros that were not doing much