feat: PG + Goblin (#4399)
Implements a new `ClientIvc` class that serves as the interface for
doing IVC with PG + Goblin. Adds a PoC for the architecture in new test
file. This test (and an updated one for Goblin + vanilla recursion) use
a new goblin method `merge` which is similar to the `accumulate` method
but with the IVC prover related work stripped out. This allows goblin to
be completely agnostic to the IVC scheme (recursion/folding) that it's
being used with.

---------

Co-authored-by: codygunton <codygunton@gmail.com>