feat: simplify public input copy cycles (#4753)
Our public input mechanism involves copying the PI into the first two
wires. The third and fourth are zero over this range. Historically, each
of these zeros has been in its own copy cycle. This change makes it so
that all of those zeros are copy constrained together. Neither is
inherently better but the latter allows the removal of special treatment
for the PI when constructing the copy cycles. This change effects the
sigma polynomials and thus changes the VK. I've updated the
VK-consistency test in the join-split suite accordingly. (I'm making
this small change in isolation because I rely on the consistency of the
VK in some of my larger PRs in this series to give the reviewer
confidence that nothing has changed unexpectedly).