refactor: Move remaining data out of Honk UltraComposer (#4848)
As the name says, move all data out of the composer. This ripples into
changes in where the commitment key, crs factory, and srs are stored and
how they are initialized. The basic idea is to use the global versions
of these things when available, and to store and initialize them early,
in either flavor classes or instance classes (which will become flavor
classes).