chore: squash yp ypb + other build improvements. (#4901)
* Get rid of `boxes-files` job as unbox command was removed from
aztec-cli.
* Get rid of `yarn-project-base`. Attempting to skip dependency
downloads was a bit of a false economy and was quite a bit of
complication.
* `ecr_login` more explicitly, rather than every time we call
`extract_repo`.
* We flatten the yarn-project jobs a bit. This means we do repeat work
(bootstrapping), but can better parallelise jobs.
* Remove all `.dockerignore` files from the repo, and ignore them from
git. Instead we compute the file from git when building in CI or using
`bootstrap_docker.sh`. This is much less error prone, and means when
testing building images locally we're getting the same as in CI. New
script is `create_docker_ignore`.
* Add `--platform=linux/amd` in some `FROM` directives where arch
doesn't matter, to improve fast bootstrap on arm systems.
* `build_local` logs into ecr just once.
* Add `[ci no-term <repo>]` option which allows connection to aws build
instances for debugging. (They still get reaped after 20m).
* Remove `l1-contracts` installing submodule deps in dockerfile. It's
expected this to have been done before building the image by the
build-system.
* Remove `end-to-end/scripts/docker-compose-browser.yml` and start
script, seems this was a workaround from some alpine build stuff, but we
just build using ubuntu now.
* Remove `unlimited` from number of jobs. Having this actually crashed
my mac when running the tests. By default we still use half the cores,
which is fine as each test suite runs tests in parallel as well.