refactor: share code between provers (#4655)
A lot of code is repeated between the Decider prover, folding prover,
and ultra honk prover. This PR aims to reduce duplication by creating a
OinkProver, which supports the 5 round functions before sumcheck. 
The OinkProver is used in the folding and ultra honk provers.

It does not address the shared code between verifiers or the shared code
between prover and verifier. It also is an initial step at a round
abstraction, where each round is implemented as a separate class and the
data being used/modified in each round is clearly defined.

Resolves https://github.com/AztecProtocol/barretenberg/issues/795.