fix: Simplify ECCVM prover constructor and add a TODO (#5681)
Simplify ECCVM prover constructor and clarify the confusion around what
we thought was the use of uninitialized polynomials via an issue and a
TODO.

Note: I think there may have been a merge mistake or something in Cody's
recent PR (which I approved) that left the ECCVM prover constructors in
a very weird state. (Also there was a WORKTODO left in code). There is
still some weirdness (captured in the TODO) but hopefully this clarifies
the situation a bit.

`ClientIVCBench/Full/6      22891 ms        17769 ms            1`