feat: naive structured execution trace (#5853)
Adds logic for executing Honk/PG with a structured execution trace, i.e.
one where each block (gate type) has a static fixed size. (This size is
arbitrarily set to 2^10 for now). This is achieved by simply offsetting
the placement of the gate data into the polynomials based on the fixed
block size in `ExecutionTrace_` rather than densely packing the blocks.

The logic is tested for full proof construction/verification and
folding/deciding for both UH and UGH. This includes a PG test that
demonstrates that the structured trace allows for folding
"inhomogeneous" circuits, i.e. circuits with a differing number of
constraints.

Note: This is not the end of the story since without additional
optimizations this approach would be extremely inefficient in terms of
memory and computation. Some of these optimizations are described in the
corresponding Milestone.

`ClientIVCBench/Full/6      21623 ms        16526 ms            1`