docs: sumcheck documentation (#5841)
Re-organized docs for pow, sumcheck, and sumcheck_round. 

- Added detailed Sumcheck Outline which contains the ZK techniques to be implemented 
- SumcheckProver/Verifier, SumcheckProver/VerifierRound got detailed
descriptions, where steps in the implementations match mathematical
descriptions. Detailed descriptions of main methods.
- PowPolynomial struct: cleaned up and moved confusing parts explaining
the computation of round univariates to the detailed description of
SumcheckProver.
- General: made notation in docs consistent with the objects in the
code, got rid of several inconsistencies, mostly in SumcheckVerifier and
SumcheckVerifierRound.

Remark1. Made docs easy to find via search in doxygen. E.g. one could
type PowPolynomial and go to the page where it's documented.
Remark2. In the case of pow, since it is a struct, the docs have to be
placed after the the code to be rendered correctly by doxygen

---------

Co-authored-by: maramihali <mara@aztecprotocol.com>