feat: representation of a grumpkin verifier commitment key inside a bn254 circuit (#6593)
This PR creates a representation of Grumpkin `VerifierCommitmentKey`
inside a bn254 circuit and ensure consistency with native part. This is
necessary to enable IPA verification in the recursive setting. I took a
minimalist approach and excluded the `pippenger_runtime_state` as we
will initially do MSMs in the IPA verifier via `batch_mul`. This is
because pippenger needs to be rewritten to be usable in a recursive
settings. I also restricted access to the `VerifierSrs` via member
functions in the native `VerifierCommitmentKey`. This work is a step
towards the ECCVM recursive verifier.