feat: consistent handling of point at infinity in transcript (#7709)
Have a fixed representation of point at infinity and engineer the
transcript to add 0s when encountering a point at infinity in order to
avoid discrepancy in challenge generation between the native and
recursive world. This is preliminary work to a (finally!) robust
handling of point at infinity. Added TODO to validate points retrieved
from the transcript when a more optimal validate_on_curve is implemented.