feat: native world state (#7516)
This PR adds a new CPP module for managing the merkle trees that make up
the world state. A new implementetion of `MerkleTreeDb` is provided to
interact with the native code.

msgpack is used to pass messages across the js<->cpp boundary. Tests
have been added to assert that the two world state implementations work
in the same way (more tests would be better).

This PR builds on top of #7037.

PS: I'm not as experienced with C++

---------

Co-authored-by: PhilWindle <philip.windle@gmail.com>