feat: new per-enqueued-call gas limit (#9033)
It is not trivial to implement "startup conditions" in the AVM circuit
that are not standard part of any execution. So, for now we think it'd
be much easier to enforce our "max gas" in the kernel. The simulator and
witgen will check the startup gas and error if its too much, but the AVM
circuit will not include constraints for this.

Note that this is an important constraint (whether in the kernel or AVM)
because for now this gas limit also serves to ensure that the AVM
circuit's trace never fills up.