feat: translator on Shplemini (#9329)
In this PR:
* implement concatenation trick (to work for both Gemini and Shplemini),
try to document it and fix some other documentation in Shplemini
* switch Translator to Shplemini

The Translator VM works on many many small polynomials (whose length is
determined by a "minicircuit size"). To avoid the permutation relation
having a very high degree, these small polynomials are split into
groups, and each group is concatenated into a single polynomial. We want
the prover to avoid having to commit to these extra concatenation
polynomials (as they will likely not be sparse at all) but rather reuse
the commitments to the polynomials in its corresponding concatenation
group, also showing they are correctly related in the opening protocol.
Briefly, in Gemini, this is achieved by adding the contributibution to
the batched concatenated polynomials when computing the fold polynomials
(A_0, A_1, ..., A_(logn -1)) but computing A_0- and A_0+ using the
polyinomials in the batched concatenated groups. As the verifier only
receives commitments to A_1, .., A_(logn-1) and has to compute the
commitments to A_0- and A_0+ , it can then do this using the commitments
of the polynomials in concatenation groups.