feat: track active accumulator rows and leverage in IVC folding (#9599)
If the IVC accumulator does not make use of the full capacity of a given
block, there will be some number of unused/empty rows at the end of the
block. We have a mechanism to skip performing relation work at such rows
but we sometimes still perform relatively expensive operations like
`get_row()` on them. This additional overhead can become very
significant if the total accumulator content is much smaller than the
structured trace size. We can avoid a lot of unnecessary work on these
empty rows by tracking the "active ranges" of the execution trace, i.e.
the regions of the accumulator which correspond to non-zero relation
contributions.

This PR introduces a class `ExecutionTraceUsageTracker` for tracking the
active regions of the accumulator and computing efficient distribution
of the execution trace rows according to actual content in the
multithreading context. This logic is leveraged in the combiner and
perturbator computations but can potentially be introduced in other
places as well.

Some high level numbers for the current benchmark case but with the 2^19
circuit replaced with another 2^17 circuit (i.e. ~2^17 worth of content
in a 2^19 ambient trace)

Master:
```
ClientIVCBench/Full/6      27137 ms        23345 ms

compute_combiner(t)                     5826   22.12%
compute_perturbator(t)                  2295    8.71%
```

Branch:
```
ClientIVCBench/Full/6      23247 ms        21532 ms

compute_combiner(t)                     2804   12.49%
compute_perturbator(t)                  1522    6.78%
```