feat: IPA accumulators setup for Rollup (#10040)
Creates a new flavor, UltraRollupFlavor, that handles IPA accumulators.
Currently unused in the rollup, but will be used there.

Adds IPA claim to builder, pk, vk, so that the verifier knows where to
look to extract the IPA claim from. Modifies the UltraRecursiveVerifier
to extract out the IPA claim from the public inputs and return it.

Also modifies native verifier to check the IPA claim and proof.