chore: Optimise grand product computation round based on active ranges (#10460)
Skip computation of numerator and denominator of z_perm at indexes that
are not part of the active ranges, and refine the threshold in
commit_structured for `z_perm`. New benchmarks:
**Now: 5.6 s difference**
We still see a difference between committing to z_perm between an
ambient trace of 2^19 and 2^20, caused by the fact that the active
ranges complement are larger (i.e. the ranges in the trace blocks where
z_perm is constant) because the blocks themselves are larger. We 
make sure to at least avoid computing and committing to z_perm after the
final active wire index.